[client] | feat(App): 페이지 마다 다르게 render 하기 위한 로직 작성
비디오 업로드 시 발생하는 Context를 비디오 업로드 페이지에서만 제공하기 위한 로직과 서버사이드에서 로그인 되어 있는지 판별 할 수 있는 로직을 작성했습니다.
