+++ packages/client/views/App/interfaces/Token.ts
@@ -1,0 +1,10 @@
+import { User } from './User';
+
+export interface Token {
+  data: {
+    sessionId: string;
+    userPublicInfo: User;
+  };
+  iat: number;
+  exp: number;
+}
