+++ packages/client/views/App/interfaces/User.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: number;
+  email: string;
+  username: string;
+  avatar: string;
+}
